requirements.txt
scipy
eyediagram
numpy
rasterio
scikit-image
shapely
tqdm
torch
torchvision
torchaudio
vit_pytorch
pillow
